% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{crt_bound}
\alias{crt_bound}
\title{#############################################################
crt_bound}
\usage{
crt_bound(
  df,
  pu,
  var_name_pu_id,
  var_name_nextdown,
  var_name_distance_length,
  connection_limit = 10,
  distance_from_meters_to_km = FALSE,
  boundary_sqrt_adjust = TRUE,
  crt_output_file = TRUE,
  outdir = getwd()
)
}
\arguments{
\item{df}{dataframe which column names include the
var_name_pu_id, var_name_nextdown, var_name_distance_length
variable names}

\item{pu}{object returned by crt_pu() or the pu file read
as dataframe. Used to validate the ids in bound file}

\item{var_name_pu_id}{character name in df that identifies
the data on the planning unit identification number}

\item{var_name_nextdown}{character name in df that
identifies the nextdown identification number}

\item{var_name_distance_length}{character name in df that
identifies the distance (length) between planning units
this is, between pu and nextdown_pu}

\item{connection_limit}{numeric; characterize stream
connectivity until a maximum of connection_limit number of
connections in the flow of the river network Default = 10}

\item{distance_from_meters_to_km}{boolean, whether to
convert distance length from meters to kilometers. Makes
sense only if distance data is in meters. Default = FALSE}

\item{boundary_sqrt_adjust}{adjust the magnitude of boundary
by dividing its value by its square root. Recommended
Default TRUE}

\item{crt_output_file}{booolean, whether to write the output
to a csv file default TRUE}

\item{outdir}{directory to write the output pu.csv file
default = getwd(), currently directory

OUTPUT}
}
\value{
pu data frame
data format (column ids): id1, id2, boundary
}
\description{
DESCRIPTION
create the boundary length file (bound.csv) which contains
information on the effective length of shared boundaries
between planning units (connectivity). The value of the
variable "boundary" is essentially a relative measure of
how important it is to include one planning unit in the
reserve system, given the inclusion of the other.
The function depends on the R package "riverPlanningTools"
developed by José Salgado-Rojas (Virgilio's group)
pkg link:
https://github.com/josesalgr/riverPlanningTools/tree/v0.1
}
\details{
PARAMETERS
}
