% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{crt_puvspr}
\alias{crt_puvspr}
\title{#############################################################
crt_puvspr}
\usage{
crt_puvspr(
  df,
  species_names,
  predictions_colnames,
  pu_id,
  crt_output_file = TRUE,
  outdir = getwd()
)
}
\arguments{
\item{df}{dataframe containing in its column names the
species names and the variable names of their predicted
occurrences/abundances per planning unit id}

\item{species_names}{vector with the names of the species}

\item{predictions_colnames}{vector of variable names in df
identifying the predicted occurrences / abundances.
predictions_colnames needs to have the same length and
follow the same order as species_names. E.g.,
if species_names     = c("sp1",      "sp2"),
predictions_colnames = c("pred_sp1", "pred_sp2")}

\item{pu_id}{the planning unit identification number
(vector)}

\item{crt_output_file}{whether to write the output to a csv
file default TRUE}

\item{outdir}{directory to write the output pu.csv file
default = getwd(), currently directory

OUTPUT}
}
\value{
pu data frame
data format (column ids): species, pu, amount
}
\description{
DESCRIPTION
create the planning unit vs conservation feature input file,
called "puvspr.csv". The function returns the puvspr data
frame object, and writes the result to outdir directory if
requested by the user (default = TRUE)
}
\details{
PARAMETERS
}
