% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{crt_spec}
\alias{crt_spec}
\title{#############################################################
crt_spec}
\usage{
crt_spec(
  puvspr,
  species_names,
  percent_target = 0.25,
  penalty = 100,
  crt_output_file = TRUE,
  outdir = getwd()
)
}
\arguments{
\item{puvspr}{object returned by crt_puvspr() or the puvspr
file read as dataframe}

\item{species_names}{vector with the names of the species}

\item{percent_target}{vector, target percent coverage per
species (in proportions, i.e., in range (0, 1)). Its length
must be equal to the length of species names. If single
value, then all species have the same target. Default = 0.25}

\item{penalty}{vector species penalty factor spf.
Penalization for not achieving the target. Must be of length
equal to species names (i.e., one spf value per species). If
single value, then the same penalization is assumed for all
species. Default = 100}

\item{crt_output_file}{whether to write the output to a csv
file default TRUE}

\item{outdir}{directory to write the output pu.csv file
default = getwd(), currently directory

OUTPUT}
}
\value{
pu data frame
data format (column ids): id, target, name, spf
}
\description{
DESCRIPTION
create the conservation feature input file, called
"spec.csv". The function returns the spec data
frame object, and writes the result to outdir directory if
requested by the user (default = TRUE)
the spec file contains information about each of the
conservation features being considered, such as their name,
target amount, and the penalty if the target is not met.
}
\details{
PARAMETERS
}
