% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_basin.r
\name{delineate_basin}
\alias{delineate_basin}
\title{#############################################################
DELINEATE BASIN	DEPRECATED}
\usage{
delineate_basin(
  reference_layer = NULL,
  polygons_ids = "HYBAS_ID",
  country = NULL,
  extent = NULL,
  mask = NULL,
  sites = NULL,
  rivers = NULL,
  lakes = NULL,
  outfile = FALSE,
  download_dir = tempdir(),
  aggregate_polygons = TRUE
)
}
\arguments{
\item{reference_layer}{(spatial vector) Reference layer
on which to draw by clicking the subset layer area of interest.
If not provided, a hydroshed basin of level lev10 is used
Default = NULL}

\item{polygons_ids}{(character) variable name
identifier that contains the polygons IDs in reference_layer.
If not provided, it takes the value used by hydrosheds to
identify basin polygons
Default "HYBAS_ID"}

\item{country}{(Char) country name of watershed location
Default = NULL}

\item{extent}{(vector) extent of interest with format:
(xmin, xmax, ymin, ymax). If an extent is provided, country
will be omitted
Default = NULL}

\item{mask}{(raster / spatial vector) reference map layer
depicting the watershed
Default = NULL}

\item{sites}{(spatial vector) of sampling sites. This
optional item can help with the delineation of the basin
area of interest
Default = FALSE}

\item{rivers}{(spatial vector) custom river network to use
instead of the hydrosheds rivers used by the package
Default = FALSE}

\item{lakes}{(spatial vector) custom lakes polygons to use
instead of the hydrosheds lakes used by the package
Default = FALSE}

\item{outfile}{(Char) filename where to write the
output delineated basin as SpatVector (make sure that
filename has the appropriate extension, e.g., *.gpkg)
Default = FALSE}

\item{download_dir}{(Char) path directory where to store
downloaded resources. This will be used to get missing
GIS data layers from a repository
Default = a temporary directory}

\item{aggregate_polygons}{(boolean) if TRUE, a spatial
vector with a single polygon describing the area of
interest is returned
Default = TRUE

OUTPUT}
}
\value{
SpatVector of basin area of interest

DEPRECATED
}
\description{
DESCRIPTION
function to interactively delineate a basin or modeling
region of interest. Returns a spatial vector made out of the
selected polygon/s. If a country is not provided, then a
valid extent or reference mask is required.
}
\details{
PARAMETERS
}
