% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_basin.r
\name{draw_by_clicking}
\alias{draw_by_clicking}
\title{#############################################################
DRAW BY CLICKING}
\usage{
draw_by_clicking(
  layer,
  IDs,
  aggregate_polygons = TRUE,
  outfile = FALSE,
  selectioncolor = "darkseagreen4"
)
}
\arguments{
\item{layer}{(spatial vector) Reference layer
on which to draw by clicking the subset layer area of interest.
Required}

\item{IDs}{(character) variable name
identifier that contains the polygons IDs in the reference
layer.
Default Required}

\item{aggregate_polygons}{(boolean) if TRUE, a spatial
vector with a single polygon describing the area of
interest is returned
Default = TRUE}

\item{outfile}{(Char) filename where to write the
output delineated basin as SpatVector (make sure that
filename has the appropriate extension, e.g., *.gpkg)
Default = FALSE}

\item{selectioncolor}{color value to highlight selection

OUTPUT}
}
\value{
(spatvector) selected subset
}
\description{
DESCRIPTION
(internal) draw selection by clicking on plotted reference
layer. Return the selected subset vector
}
\details{
PARAMETERS
}
