% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_basin.r
\name{draw_subset_by_clicking}
\alias{draw_subset_by_clicking}
\title{#############################################################
DRAW SUBSET BY CLICKING}
\usage{
draw_subset_by_clicking(
  x,
  polygons_ids,
  map = NULL,
  sites = NULL,
  rivers = NULL,
  lakes = NULL,
  outfile = FALSE,
  aggregate_polygons = TRUE
)
}
\arguments{
\item{x}{(spatial vector) target spatvector layer from which
to extract a subset by clicking on a plot.
Required}

\item{polygons_ids}{(character) variable name in x that
identifies its polygons.
Required}

\item{map}{(spatvector) reference layer: country borders
Default NULL}

\item{sites}{(spatvector) reference layer: sampling locations
Default = NULL}

\item{rivers}{(spatvector) reference layer: river network
Default = NULL}

\item{lakes}{(spatvector) reference layer: lakes
Default = NULL}

\item{outfile}{(Char) filename where to write the subset
output (make sure that the file name has the appropriate
extension, e.g., *.gpkg)
Default = FALSE}

\item{aggregate_polygons}{(boolean) if TRUE, merge polygons
that share borders
Default = TRUE

OUTPUT}
}
\value{
spatvector subset
}
\description{
DESCRIPTION
function to interactively select a subset of polygons.
Returns a spatial vector made out of the
selected polygon/s. Reference layers can be optionally
provided.
}
\details{
PARAMETERS
}
