% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrolinked_data.r
\name{extract_hydrolinked_data}
\alias{extract_hydrolinked_data}
\title{#############################################################
EXTRACT HYDROLINKED DATA}
\usage{
extract_hydrolinked_data(x, predictor_raster, fun = "sum", ...)
}
\arguments{
\item{x}{(SpatVector) as given by
marxan.toolbox::get_rivnet_SC. If another SpatVector of
polygons is used, make sure that it has its hydrological
connectivity stored in a variable called "upstream"
REQUIRED}

\item{fun}{(character) function to summarize the extracted
data
DEFAULT "sum"}

\item{...}{further arguments passed to terra::extract

OUTPUT}

\item{predictor_layer}{(SpatRaster or SpatRaster stack)
from which to extract the data
REQUIRED}
}
\value{
River object. Information details in Rivnet package
documentation
}
\description{
DESCRIPTION
method to extract data based on the hydrological connection.
The function makes use of the polygons ids stored in the
upstream variable (obtained after applying
marxan.toolbox::get_rivnet_SC) to build the upstream
connectivity. By default, cummulative predictor values
are extracted (i.e., fun = "sum")
}
\details{
PARAMETERS
}
