% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{get_best_solution}
\alias{get_best_solution}
\title{#############################################################
GET BEST (MARXAN) SOLUTION
DESCRIPTION
returns the best marxan solution located in
marxan_dir/output/output_best.csv
PARAMETERS}
\usage{
get_best_solution(marxan_dir = getwd())
}
\arguments{
\item{marxan_dir}{marxan directory (where input.dat file is
located)

OUTPUT}
}
\value{
(vector) with best marxan solution. A value of 1
means that that planning unit was selected as part of the
solution by the optimization algorithm. Zero otherwise
}
\description{
#############################################################
GET BEST (MARXAN) SOLUTION
DESCRIPTION
returns the best marxan solution located in
marxan_dir/output/output_best.csv
PARAMETERS
}
