% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_basin.r
\name{get_gdrive_resource}
\alias{get_gdrive_resource}
\title{#############################################################
GET GDRIVE RESOURCE}
\usage{
get_gdrive_resource(id, temp_dir = FALSE)
}
\arguments{
\item{id}{(Char) google drive file identification
REQUIRED}

\item{temp_dir}{(Char) path directory where to store
downloaded resources
Default = FALSE i.e., a temporary directory

OUTPUT}
}
\value{
downdloaded filename
}
\description{
DESCRIPTION
(internal) download google drive file by id and returns the
unzipped filename
}
\details{
PARAMETERS
}
