% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivnet_utils.r
\name{get_rivnet_SC}
\alias{get_rivnet_SC}
\title{#############################################################
GET RIVER NETWORK AND SUBCATCHMENTS AS SPATVECTORS}
\usage{
get_rivnet_SC(
  river,
  SC_res = 20,
  thrA_factor = 0.002,
  crs = "epsg:3035",
  outdir = ".",
  tag = "",
  collection = FALSE
)
}
\arguments{
\item{river}{(river object) as given by
marxan.toolbox::rivnet_extract_river
REQUIRED}

\item{thrA_factor}{(numeric double) scaling factor to
thrA parameter in OCNet::aggreate_OCN: threshold value on
drainage area used to derive the aggregated network.
If thrA = 0, no aggregation is performed
DEFAULT = 0.002}

\item{crs}{(character) coordinate reference system used
when creating the SC layer. It is recommended to input the
same crs used for the extraction of the river network}

\item{outdir}{(character) output directory where to store
the river and SC shape files
DEFAULT CURRENT WORKING DIRECTORY}

\item{tag}{(character) custom tag to add to the ouput river
and SC shape files. Eg., catchment id, country, ...
DEFAULT "" (empty character)}

\item{collection}{(boolean) if TRUE returns the subcatchments
SC and rivernetwork as a spatvector collection. Otherwise,
returns the SC spatvector object only
DEFAULT FALSE

OUTPUT}

\item{SC_red}{(positive integer) scaling factor affecting
the resolution or area size of the subcatchment polygons
Default = 20}
}
\value{
River object. Information details in Rivnet package
documentation
}
\description{
DESCRIPTION
function to obtain the river network and its subcatchment
polygons as spatvectors from a river object (as returned
by marxan.toolbox::rivnet_extract_river)
}
\details{
PARAMETERS
}
