% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{knee}
\alias{knee}
\title{#############################################################
KNEE DECISION METHOD}
\usage{
knee(y, x = seq(from = 1, to = length(y), by = 1))
}
\arguments{
\item{y}{vector of length > 2. E.g., in the context of
this marxan toolbox package, the cost values}

\item{x}{vector of the same length as y. E.g., in the
context of this marxan toolbox package, the boundary or
connectivity values

OUTPUT}
}
\value{
the index where the optimal value is located
}
\description{
DESCRIPTION
Developed according to the Knee method described in
Coffey et al (2019)
The knee is defined as the intersection point of two lines
that minimizes the SSE (sum of squared errors)
}
\details{
PARAMETERS
}
