% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{next_ten}
\alias{next_ten}
\title{#############################################################
NEXT TEN
DESCRIPTION
returns the next ten of the number x
If x is a vector v, then x = max(v)
PARAMETERS}
\usage{
next_ten(x)
}
\arguments{
\item{x}{numeric or vector

OUTPUT}
}
\value{
the next ten of x (e.g., if x = 41, then return 50)
}
\description{
#############################################################
NEXT TEN
DESCRIPTION
returns the next ten of the number x
If x is a vector v, then x = max(v)
PARAMETERS
}
