% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivnet_utils.r
\name{rivnet_extract_river}
\alias{rivnet_extract_river}
\title{#############################################################
RIVNET EXTRACT RIVER}
\usage{
rivnet_extract_river(
  dem_fname,
  outlet = 1,
  locations = NULL,
  country_map = NULL,
  reference_river = NULL,
  crt_outlets_log = FALSE,
  ...
)
}
\arguments{
\item{dem_fname}{(character) file name of the digital
elevation model DEM of the catchment of interest. The DEM
should embrace all rivers upstream of the selected outlet.
If streams are cut by the DEM, the method can fail to
delineate the river network. In addition, the DEM projection
should be in meters (e.g., the european epsg:3035)
REQUIRED}

\item{outlet}{(vector, matrix, or data frame) if vector of
length 1, the user will be asked to interactively
draw the outlet(s) on the DEM map. If vector length > 1,
the odd components identify the longitudinal (x)
coordinates, and the even components the latitudinal (y)
coordinates. If a matrix, it should have 2 columns
(for x and y coordinates respectively) and number of rows
equal to the number of outlets. If a data frame, it should
have components 'outlet$x', 'outlet$y' identifying the
respective coordinates.
Default = 1}

\item{locations}{(spatvector) optional. To display point
locations of interest to use as a reference when drawing
the outlet(s)
DEFAULT NULL}

\item{country_map}{(spatvector) optional. Display country
borders when manually selecting the outlets
DEFAULT NULL}

\item{reference_river}{(spatVector) optional. To help
manually setting the outlets using a reference river layer}

\item{crt_outlets_log}{(boolean) whether to write the selected
outlets to a file
DEFAULT FALSE}

\item{...}{further arguments passed to
rivnet::extract_river

OUTPUT}
}
\value{
River object. Information details in Rivnet package
documentation
}
\description{
DESCRIPTION
function to extract the river network associated to the
provided DEM and outlet(s), using the "extract_river"
method of package "rivnet"
}
\details{
PARAMETERS
}
