% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.toolbox.r
\name{run_marxan}
\alias{run_marxan}
\title{#############################################################
RUN MARXAN (FROM R)
DESCRIPTION
run the marxan optimization algorithm from R
PARAMETERS}
\usage{
run_marxan(marxan_dir = getwd())
}
\arguments{
\item{marxan_dir}{marxan directory (where input.dat file is
located)

OUTPUT
None}
}
\description{
#############################################################
RUN MARXAN (FROM R)
DESCRIPTION
run the marxan optimization algorithm from R
PARAMETERS
}
