% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivnet_utils.r
\name{select_outlets_from_map}
\alias{select_outlets_from_map}
\title{#############################################################
SELECT OUTLETS FROM MAP}
\usage{
select_outlets_from_map(
  n_outlets = 1,
  dem,
  sites = NULL,
  country_map = NULL,
  reference_river = NULL
)
}
\arguments{
\item{n_outlets}{(numeric) requested outlet points to draw
Default = 1}

\item{dem}{(spatraster) DEM layer to draw the outlets on
REQUIRED}

\item{sites}{(spatvector) reference locations (e.g.,
sampling sites)
DEFAULT NULL}

\item{country_map}{(spatvector) reference map with country
borders
DEFAULT NULL}

\item{reference_river}{(spatvector) reference river layer
DEFAULT NULL

OUTPUT}
}
\value{
n outlets with geographic coordinates
}
\description{
DESCRIPTION
function to select n outlets by drawing points interactively
on a given DEM
}
\details{
PARAMETERS
}
