% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate_basin.r
\name{delineate_basin}
\alias{delineate_basin}
\title{#############################################################
DELINEATE BASIN}
\usage{
delineate_basin(
  country = NULL,
  extent = NULL,
  mask = NULL,
  sites = NULL,
  rivers = NULL,
  lakes = NULL,
  basin = NULL,
  outfile = FALSE,
  aggregate_polygons = TRUE
)
}
\arguments{
\item{country}{(Char) country name of watershed location
Default = NULL}

\item{extent}{(vector) extent of interest with format:
(xmin, xmax, ymin, ymax). If an extent is provided, country
will be omitted
Default = NULL}

\item{mask}{(raster / spatial vector) reference map layer
depicting the watershed
Default = NULL}

\item{sites}{(spatial vector) of sampling sites. This
optional item can help with the delineation of the basin
area of interest
Default = FALSE}

\item{rivers}{(spatial vector) custom river network to use
instead of the hydrosheds rivers used by the package
Default = FALSE}

\item{lakes}{(spatial vector) custom lakes polygons to use
instead of the hydrosheds lakes used by the package
Default = FALSE}

\item{basin}{(spatial vector) hydroshed basin of level other
than lev10 which is the one used by the package
Default = FALSE}

\item{outfile}{(character) filename where to write the
output delineated basin as SpatVector (make sure that
filename has the appropriate extension, e.g., *.gpkg)
Default = FALSE}

\item{aggregate_polygons}{(boolean) if TRUE, a spatial
vector with a single polygon describing the basin area of
interest is produced
Default = TRUE

OUTPUT}
}
\value{
SpatVector of basin area of interest
}
\description{
DESCRIPTION
function to interactively delineate a basin. Returns a
spatial vector with polygon/s representing the basin area of
interest. If a country is not provided, then a valid extent
or reference mask is required.
}
\details{
PARAMETERS
}
