% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEM.r
\name{getDEM}
\alias{getDEM}
\title{#############################################################
GET DIGITAL ELEVATION MODEL}
\usage{
getDEM(
  raster_template,
  zoom_level = 11,
  country_map = NULL,
  sites = NULL,
  outfile = FALSE,
  draw_extent = TRUE,
  ...
)
}
\arguments{
\item{raster_template}{(spatraster) spatraster object that
contains the region of interest from which to obtain the DEM
REQUIRED}

\item{zoom_level}{(ranges from 1 to 14) affects the
resolution of the DEM. More details at
\url{https://github.com/tilezen/joerd/blob/master/docs/data-sources.md#what-is-the-ground-resolution}}

\item{country_map}{(spatvector) optional. To display country
borders when manually selecting the extent of interest
Default = NULL}

\item{sites}{(spatvector) optional. To display point
locations/sites of interest to use as a reference when drawing
the extent of interest
DEFAULT NULL}

\item{outfile}{if not FALSE, write the DEM to this output file
DEFAULT = FALSE}

\item{draw_extent}{(boolean) whether to define the extent of
interest by drawing it directly on the map. The default value
is set to TRUE, because depending on the resolution the
resulting DEM could require a large disk space
DEFAULT = TRUE}

\item{...}{further arguments passed to
elevatr::get_elev_raster

OUTPUT}
}
\value{
spatraster object (DEM)
}
\description{
DESCRIPTION
function to obtain a digital elevation model or DEM from
the Amazon Web Services Terrain Tiles and the Open Topography
global datasets API,
by using the method get_elev_raster of package elevatr.
}
\details{
PARAMETERS
}
